Finish resilient SEO metadata generation
